requirements.txt
ultralytics~=8.3.89
streamlit
opencv-python~=4.11.0.86
Pillow
numpy~=2.1.1
roboflow
torch~=2.6.0
requests~=2.32.3
flask~=3.1.2
python-dotenv
paho-mqtt~=2.1.0